different.tan

different.tan: R code for the analysis of vehicle emissions remote sensing system (VERSS) data

different.tan

different.tan was initially developed as part of work on a series of projects collecting and analyzing EDAR data, but many also be of interest to those using other similar (VERSS) methods.

The name 'different.tan' refers to Don Stedman, a pioneer in field, who died in 2016: [DON]

References

We gratefully acknowledge the huge contributions of the R Core Team and numerous contributors in developing and maintaining R:

R Development Core Team (2011). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. ISBN 3-900051-07-0, URL http://www.R-project.org/.

Our own related work includes:

Ropkins, K., DeFries, T.H., Pope, F., Green, D.C., Kemper, J., Kishan, S., Fuller, G.W., Li, H., Sidebottom, J., Crilley, L.R. and Kramer, L. (2017). Evaluation of EDAR vehicle emissions remote sensing technology. Science of the Total Environment, 609, pp.1464-1474. https://doi.org/10.1016/j.scitotenv.2017.07.137

Ghaffarpasand, O., Beddows, D.C., Ropkins, K. and Pope, F.D. (2020). Real-world assessment of vehicle air pollutant emissions subset by vehicle type, fuel and EURO class: New findings from the recent UK EDAR field campaigns, and implications for emissions restricted zones. Science of The Total Environment, 734, p.139416. https://doi.org/10.1016/j.scitotenv.2020.139416

Ghaffarpasand, O., Ropkins, K., Beddows, D.C. and Pope, F.D. (2022). Detecting high emitting vehicle subsets using emission remote sensing systems. Science of The Total Environment, p.159814.https://doi.org/10.1016/j.scitotenv.2022.159814

Author

Karl Ropkins, Francis Pope, David Beddows, Omid Ghaffarpasand