import_featdata

Function imports data from the Denver FEAT data archive (See below).

import_featdata(featdata = NULL, simplify.names = TRUE, ...)

Arguments

featdata

Character, data set name (see below)

simplify.names

Logical, if TRUE, simplifies the column names of the found data set by converting them to lower case and changing underscores to dots.

...

Currently ignored.

Value

import_featdata attempts to match featdata and, if found, returns it as a data.frame.

The function uses base R code and foreign::read.dbf to download, read and tidy data. The as.is=TRUE argument is applied when reading data so, for example, date and time columns are imported as character

vectors rather than factors.

Details

The function imports data from the FEAT data archive.

dataset	datatype	ref

Tucson 94	light duty vehicles	10.1021/Es950191j

Phnx 98	light duty vehicles	10.1021/es702413b

Phnx 99	light duty vehicles	10.1021/es702413b

AZ 2000	light duty vehicles	10.1021/es702413b

AZ 2002	light duty vehicles	10.1021/es702413b

AZ 2004	light duty vehicles	10.1021/es702413b

AZ 2006	light duty vehicles	10.1021/es702413b

Fresno 08	light duty vehicles	10.1021/acs.est.0c05433

Fresno 21	light duty vehicles	10.1021/acs.est.0c05433

Cal 89	light duty vehicles	10.1126/science.268.5213.991

Cal 91	light duty vehicles	10.1126/science.268.5213.991

LA710 99	light duty vehicles	-

Lynwd105 2018	light duty vehicles	10.1080/10962247.2019.1611677

Lynwd710 2018	light duty vehicles	10.1080/10962247.2019.1611677

LA 99	light duty vehicles	10.1021/es026340x

LA 00	light duty vehicles	10.1021/es026340x

LA 01	light duty vehicles	10.1021/es026340x

Smnto 99	light duty vehicles	-

Sanjose 99	light duty vehicles	10.1021/acs.est.0c05433; 10.1021/acs.est.0c05433

Sanjose 08	light duty vehicles	10.1021/acs.est.0c05433; 10.1021/acs.est.0c05433

Vannuys 10	light duty vehicles	10.1080/10962247.2012.699015; 10.1080/10962247.2012.699016

Labrea 01	light duty vehicles	10.1080/10962247.2020.1869121

Labrea 03	light duty vehicles	10.1021/es702413b

Labrea 05	light duty vehicles	10.1021/acs.est.0c05433

Labrea 08	light duty vehicles	10.1021/es5043518

Labrea 13	light duty vehicles	10.1021/acs.est.5b02392

Labrea 15	light duty vehicles	10.1021/acs.est.6b00717

Labrea 18	light duty vehicles	10.1021/acs.est.0c05433

Labrea 99	light duty vehicles	10.1080/10962247.2019.1611677

Peralta 97	heavy duty vehicles	10.1021/Es001533a

Peralta 08	heavy duty vehicles	10.1021/es202392g

Peralta 09	heavy duty vehicles	10.1021/es202392g

Peralta 10	heavy duty vehicles	10.1021/es202392g

Peralta 12	heavy duty vehicles	10.1021/es401487b

Peralta 17	heavy duty vehicles	10.1021/acs.est.8b03994

laport 08	heavy duty vehicles	10.1021/es202392g, 10.1021/es401487b

laport 09	heavy duty vehicles	10.1021/es202392g, 10.1021/es401487b

laport 10	heavy duty vehicles	10.1021/es202392g, 10.1021/es401487b

laport 12	heavy duty vehicles	10.1021/es202392g, 10.1021/es401487b

coors 99	heavy duty vehicles	10.1021/Es001533a

coors 05	heavy duty vehicles	10.1021/es060989a

dumont 99	heavy duty vehicles	10.1021/Es001533a

dumont 05	heavy duty vehicles	10.1021/es060989a

uri 98	heavy duty vehicles	10.1021/Es001533a

pohouston 09	heavy duty vehicles	-

sanmarcos 98	heavy duty vehicles	10.1021/Es001533a

utah 20	heavy duty vehicles	10.1021/Es001533a

cottonwood 13	heavy duty vehicles	10.1021/es505534e

cottonwood 15	heavy duty vehicles	10.1021/acs.est.6b06172

cottonwood 17	heavy duty vehicles	10.1021/acs.est.8b00621

pola 13	heavy duty vehicles	10.1021/es505534e

pola 15	heavy duty vehicles	10.1021/acs.est.6b06172

pola 17	heavy duty vehicles	10.1021/acs.est.8b00621

Note

In-development: code may be subject to significant change; Currently, only some data sets can be downloaded from the FEAT archive. (See Details above for list of data sets available using import_featdata; please check main archive, if you are after any other FEAT data not listed here.)

References

For foreign package and read.dbf:

R Core Team (2022). _foreign: Read Data Stored by 'Minitab', 'S', 'SAS', 'SPSS', 'Stata', 'Systat', 'Weka', 'dBase', ..._. R package version 0.8-82, <https://CRAN.R-project.org/package=foreign>.

For general information about FEAT:

D.H. Stedman and G.A. Bishop. Measuring the Emissions of Passing Cars, Acc. Chem. Res., 29:489-495, 1996.

D.A. Burgard, G.A. Bishop, R.S. Stadtmuller, T.R. Dalton and D.H. Stedman, Spectroscopy Applied to On-Road Mobile Source Emissions. Appl. Spectrosc., 60:5:135A-148A, 2006.

And, for specific campaigns/data sets, please reference using study reference(s), listed as associated DOI(s) in Details above.

See also

Examples

#to do...